Working parts uploader and shop viewer (prototype)

+++ backend/app/models/parts.py
@@ -5,8 +5,8 @@ from datetime import datetime
 @dataclass
 class Parts(db.Model):
     part_number: str = db.Column(db.String(100), nullable=False, primary_key=True)
-    price: int = db.Column(db.Integer, nullable=False)
-    quantity: int = db.Column(db.Integer, nullable=False)
+    price: float = db.Column(db.Float, nullable=False)
+    quantity: float = db.Column(db.Float, nullable=False)
     vehicle: str = db.Column(db.String(100), nullable=True)
     vehicle_make: str = db.Column(db.String(100), nullable=True)
     vehicle_model: str = db.Column(db.String(100), nullable=True)
